chatbpt-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pyjwt
python-dotenv
google-generativeai
pydantic>=2.0
pydantic-settings>=2.0